refactor<user-service>: separate entity definition
I decided to use typeorm 'separating entity definition', which allow me to create table based on previously created interface.

Thanks to this solution, my code is typesafe and more readable - besides of above-mentioned interface I also have 'redis-user.interface' relating to redis database (which is shared between microservices)and 'extended-user.interface' which join this 2 interfaces.

Anyway, using decorators is IMO the most readable and fastest way to define entity, so I decided to stay with this pattern in other microservices.

+ added function which lowercase any string (/utils/lowercase)